Crate flowrstructs

Modules

function
function defines functions that form part of a flow

input
input defines the struct for inputs to functions in a flow

lib_manifest
lib_manifest defines the structs for specifying a Library's manifest and methods to load it

manifest
manifest is the struct that specifies the manifest of functions in a flow

output_connection
output_connection defines a struct for a function's output connection

Structs

Error
The Error type.

Enums

ErrorKind
The kind of an error.

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Type Definitions

Result
Convenient wrapper around std::Result.